requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
langchain
chromadb
sentence-transformers
transformers            # optional
ollama                 # optional (client wrapper)